megamath final + some v2

+++ auctionhouse/auctionhouse/advay.py
@@ -38,7 +38,6 @@ def logistic_function(x):
 
 
 def ElDesafio(wallet, history):
-
     if len(history) == 0:
         return 0
 
@@ -46,7 +45,7 @@ def ElDesafio(wallet, history):
         return 1
 
     sum_of_wallet = sum(list(map(lambda x: x[0], history)))
-    valid_amount = sum_of_wallet * 0.50
+    valid_amount = int(sum_of_wallet * 0.50)
 
     return 0 if valid_amount > wallet else valid_amount + 1
